Ported Jupyter data generator to Java

+++ common/src/main/java/net/mcmetrics/common/analytic/player/PlayerPurchaseAnalytic.java
@@ -16,7 +16,7 @@ public class PlayerPurchaseAnalytic implements NamedAnalytic {
     private final @NotNull UUID playerUUID;
     private final @NotNull String productName;
     private final @NotNull String currency;
-    private final @NotNull Float purchaseValue;
+    private final @NotNull Double purchaseValue;
 
     @Override
     public @NotNull String getEventType() {
